3.1.0 support, still a work in progress you will need to update your dependencies located in extras/dependencies a log of the work so far follows
*WeatherMgr now iterates over map sessions rather than world sessions to send weather info by zone, also fixed a bug/typos in MapMgr::SendPacketToPlayers

Revision: 15
Author: Pepsi1x1
Date: 0:35:21, March-29-09
Message:
*Re-enabled spline, if the number of points in the spline / distance between points is less than 1.0f we will send monster_move instead
----
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
Modified : /trunk/src/hearthstone-world/Object.cpp
 
Revision: 14
Author: Chameleon
Date: 6:21:58, March-28-09
Message:
* Started LEARN_PREVIEW_TALENTS handler. Doesn't handle learning talent ranks other than current + 1 yet
* Always send full talents update for now until partial update is fixed
* Use GAMEOBJ_USE handler also for GAMEOBJ_REPORT_USE
----
Modified : /trunk/src/hearthstone-world/Player.cpp
Modified : /trunk/src/hearthstone-world/SkillHandler.cpp
Modified : /trunk/src/hearthstone-world/WorldSession.cpp
Modified : /trunk/src/hearthstone-world/WorldSession.h
 
Revision: 13
Author: Pepsi1x1
Date: 4:26:08, March-27-09
Message:
*Updated to newest client 9733
*Spline properly fixed :D
*Tweaked a9 again
*Started working on monster_move
----
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
Modified : /trunk/src/hearthstone-world/Object.cpp
Modified : /trunk/src/hearthstone-world/UpdateFields.h
 
Revision: 12
Author: Chameleon
Date: 4:11:16, March-27-09
Message:
* SMSG_TALENTS_INFO implemented. Can see the talents you have now
* INITIALIZE_FACTIONS sends 0s for now to prevent assert
* Hackfix: Don't send TARGET_FLAG_DEST_LOCATION in SPELL_GO to prevent crash
* Started handlers for SpellEffectSetTalentSpecsCount and SpellEffectActivateTalentSpec
* Disabled achievement criteria packet cause it spams a lot
----
Modified : /trunk/src/hearthstone-world/AchievementInterface.cpp
Modified : /trunk/src/hearthstone-world/Player.cpp
Modified : /trunk/src/hearthstone-world/Player.h
Modified : /trunk/src/hearthstone-world/ReputationHandler.cpp
Modified : /trunk/src/hearthstone-world/SkillHandler.cpp
Modified : /trunk/src/hearthstone-world/Spell.cpp
Modified : /trunk/src/hearthstone-world/Spell.h
Modified : /trunk/src/hearthstone-world/SpellEffects.cpp
 
Revision: 11
Author: Chameleon
Date: 5:54:50, March-26-09
Message:
BATTLEFIELD_LIST update
----
Modified : /trunk/src/hearthstone-world/BattlegroundHandler.cpp
Modified : /trunk/src/hearthstone-world/BattlegroundMgr.cpp
 
Revision: 10
Author: Chameleon
Date: 4:35:36, March-26-09
Message:
Updated query_handlers structure. Still need to make them send correct info from db in correct fields.
----
Modified : /trunk/src/hearthstone-realmserver/QueryHandler.cpp
Modified : /trunk/src/hearthstone-realmserver/Structures.h
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
 
Revision: 9
Author: Chameleon
Date: 1:05:26, March-26-09
Message:
* Updated SMSG_ACTION_BUTTONS
* Reverted number of spline points to 4 as it crashes on ASSERT(count>3)
* Disabled spline sending for now
----
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
Modified : /trunk/src/hearthstone-world/Object.cpp
Modified : /trunk/src/hearthstone-world/Player.cpp
 
Revision: 8
Author: Pepsi1x1
Date: 17:25:55, March-25-09
Message:
*Commiting code that compiles would be helpful, oops
*More work on the spline mover code, missed a uint32
-I think the count in the spline should be 3 not 4 since were sending 3 positions, gotta test this
----
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
Modified : /trunk/src/hearthstone-world/Creature.cpp
Modified : /trunk/src/hearthstone-world/Creature.h
Modified : /trunk/src/hearthstone-world/QuestMgr.cpp
Modified : /trunk/src/hearthstone-world/TransporterHandler.cpp
 
Revision: 7
Author: Pepsi1x1
Date: 13:59:05, March-25-09
Message:
*Some further work on the update packets
----
Modified : /trunk/src/hearthstone-world/Object.cpp
Modified : /trunk/src/hearthstone-world/Creature.cpp
Modified : /trunk/src/hearthstone-world/Creature.h
Modified : /trunk/src/hearthstone-world/TransporterHandler.cpp
 
Revision: 6
Author: Pepsi1x1
Date: 5:28:44, March-25-09
Message:
*crashfix for non player caster in spell start in dk rune update check
----
Modified : /trunk/src/hearthstone-world/Spell.cpp
 
Revision: 5
Author: Pepsi1x1
Date: 17:43:54, March-24-09
Message:
*Updated spellcasttargets, seems i missed the stackpacket version when i updated it for 3.0.8, should fix that spell assert NTY, need to get some sleep
----
Modified : /trunk/src/hearthstone-world/Spell.cpp
 
Revision: 4
Author: Chameleon
Date: 17:09:38, March-24-09
Message:
* update for INIT_SPELLS
* don't send rune update to not DKs
----
Modified : /trunk/src/hearthstone-world/Spell.cpp
Modified : /trunk/src/hearthstone-world/Player.cpp
 
Revision: 3
Author: Pepsi1x1
Date: 17:01:27, March-24-09
Message:
*3.1.0 Ptr In World
*Added RC4 encryption thanks to the guys at mangos
*New openssl lib
*Change GenerateKey to accept a seed
*New Opcodes
*Extracted new update fields
*Glyphs now only have 6 fields, apparantly these are not done in the a9 anymore havent changed this though
*Updated a9 struct, not sure about vehicles
*Updated a9 update flags, need to double check these
*Updated SMSG_MONSTER_MOVE struct
*Updated spline struct
*Updated SMSG_DESTROY_OBJECT missed this when moving from AC to HS in 3.0 update
*Tidied up coding style in ProcessPendingUpdates was a proper mess
----
Modified : /trunk/src/hearthstone-world/Player.cpp
Modified : /trunk/extras/dependencies/DLL Files/libeay32.dll
Modified : /trunk/extras/dependencies/DLL Files (x64)/libeay32.dll
Modified : /trunk/extras/dependencies/VC/lib/amd64/libeay32.lib
Modified : /trunk/extras/dependencies/VC/lib/libeay32.lib
Modified : /trunk/src/hearthstone-realmserver/WorldSocket.cpp
Modified : /trunk/src/hearthstone-shared/Auth/AuthPacketKey.h
Added : /trunk/src/hearthstone-shared/Auth/SARC4.cpp
Added : /trunk/src/hearthstone-shared/Auth/SARC4.h
Modified : /trunk/src/hearthstone-shared/Auth/WowCrypt.cpp
Modified : /trunk/src/hearthstone-shared/Auth/WowCrypt.h
Modified : /trunk/src/hearthstone-shared/RC4Engine.h
Modified : /trunk/src/hearthstone-world/AIInterface.cpp
Modified : /trunk/src/hearthstone-world/Corpse.cpp
Modified : /trunk/src/hearthstone-world/EyeOfTheStorm.cpp
Modified : /trunk/src/hearthstone-world/GameObject.cpp
Modified : /trunk/src/hearthstone-world/Item.cpp
Modified : /trunk/src/hearthstone-world/ItemInterface.cpp
Modified : /trunk/src/hearthstone-world/Level3.cpp
Modified : /trunk/src/hearthstone-world/LogonCommClient.cpp
Modified : /trunk/src/hearthstone-world/Object.cpp
Modified : /trunk/src/hearthstone-world/Object.h
Modified : /trunk/src/hearthstone-world/Opcodes.cpp
Modified : /trunk/src/hearthstone-world/Opcodes.h
Modified : /trunk/src/hearthstone-world/SpellEffects.cpp
Modified : /trunk/src/hearthstone-world/TaxiMgr.cpp
Modified : /trunk/src/hearthstone-world/UpdateFields.h
Modified : /trunk/src/hearthstone-world/Vehicle.cpp
Modified : /trunk/src/hearthstone-world/WarsongGulch.cpp
Modified : /trunk/src/hearthstone-world/WorldSession.cpp
Modified : /trunk/src/hearthstone-world/WorldSocket.cpp
Modified : /trunk/src/hearthstone-world/WorldSocket.h
Modified : /trunk/win/VC90/hearthstone-shared.vcproj


git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/3_1_0@1229 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Object.h
@@ -519,7 +519,7 @@ protected:
 	//! Mark values that player should get when he/she/it sees object for first time.
 	virtual void _SetCreateBits(UpdateMask *updateMask, PlayerPointer target) const;
 
-	void _BuildMovementUpdate( ByteBuffer *data, uint8 flags, uint32 flags2, PlayerPointer target );
+	void _BuildMovementUpdate( ByteBuffer *data, uint16 flags, uint32 flags2, PlayerPointer target );
 	void _BuildValuesUpdate( ByteBuffer *data, UpdateMask *updateMask, PlayerPointer target );
 
 	/* Main Function called by isInFront(); */
@@ -645,4 +645,5 @@ public:
 };
 
 #endif
+
